CM-36350, CM-36348, CM-36349, CM-36353, CM-36355 - Add scan for hardcoded secrets (#9)

+++ src/extension/Cycode.VisualStudio.Extension.Shared/Constants.cs
@@ -15,8 +15,8 @@ public static class Constants
     public const string CliExecutableAssetName = "cycode-win.exe";
     public const string CliExecutableShaAssetName = $"{CliExecutableAssetName}.sha256";
 
-    public const long PluginAutoSaveFlushInitialDelaySec = 0L;
-    public const long PluginAutoSaveFlushDelaySec = 5L;
+    public const int PluginAutoSaveFlushInitialDelaySec = 0;
+    public const int PluginAutoSaveFlushDelaySec = 5;
 
     private static string GetPluginsPath()
     {
